Refactor folder structure (move libraries to lib/src folder)
